@many to many bidirectional - starter

+++ src/main/java/com/notyfyd/service/RoleService.java
@@ -48,10 +48,12 @@ public class RoleService {
     /** Delete a specified role given the id */
     public ResponseEntity<Object> deleteRole(Long id) {
         if(roleRepository.findById(id).isPresent()){
-            roleRepository.deleteById(id);
-            if(roleRepository.findById(id).isPresent()){
-                return ResponseEntity.unprocessableEntity().body("Failed to delete the specified record");
-            } else return ResponseEntity.ok().body("Successfully deleted specified record");
+            if(roleRepository.getOne(id).getUsers().size() == 0) {
+                roleRepository.deleteById(id);
+                if (roleRepository.findById(id).isPresent()) {
+                    return ResponseEntity.unprocessableEntity().body("Failed to delete the specified record");
+                } else return ResponseEntity.ok().body("Successfully deleted specified record");
+            } else return ResponseEntity.unprocessableEntity().body("Failed to delete,  Please delete the users associated with this role");
         } else
             return ResponseEntity.unprocessableEntity().body("No Records Found");
     }
